# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "bale-broadcast-ai API",
    "version": "1.0.0",
    "description": "API واحد و نسخه‌بندی‌شده‌ی پلتفرم پیام‌رسانی/فروش هوشمند روی بله.\n\n## احراز هویت\nهمه‌ی endpoint ها (به‌جز /auth/register و /auth/login) به هدر Authorization: Bearer <token> نیاز دارند. دو نوع توکن پذیرفته می‌شود و هر دو از یک هدر یکسان استفاده می‌کنند:\n\n1. JWT نشست داشبورد — از POST /auth/login یا POST /auth/register گرفته می‌شود. اعتبار ۷ روزه دارد.\n   مثال: Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n2. کلید API بلندمدت — از POST /api-keys ساخته می‌شود، با پیشوند abai_live_ شروع می‌شود.\n   مثال: Authorization: Bearer abai_live_9f6e817cb2a1...\n\nسرور نوع توکن را خودکار از روی پیشوندش تشخیص می‌دهد. هر دو userId احراز هویت‌شده را تعیین می‌کنند و پلتفرم داده‌ی هر کاربر را کاملاً از بقیه ایزوله نگه می‌دارد.\n\n## کدهای خطای استاندارد\n- 401: توکن ارسال نشده، نامعتبر، منقضی یا باطل‌شده است.\n- 403: توکن معتبر است ولی اجازه‌ی دسترسی به این بخش را ندارد (مثلاً /admin/* فقط برای owner).\n\nهر پاسخ خطا یک شیء ساده با فیلد error (رشته‌ی قابل‌نمایش) برمی‌گرداند."
  },
  "servers": [{ "url": "/api/v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT یا API Key",
        "description": "یا JWT گرفته‌شده از /auth/login و /auth/register، یا یک کلید API که با POST /api-keys ساخته‌اید (پیشوند abai_live_). هر دو در همین یک هدر Authorization: Bearer <token> فرستاده می‌شوند."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "درخواست نامعتبر — فیلد الزامی خالی است یا مقدار داده‌شده نادرست است.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "فیلدهای name و descriptionText الزامی هستند." } } }
      },
      "Unauthorized": {
        "description": "توکن ارسال نشده، نامعتبر، منقضی یا باطل‌شده است.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "توکن احراز هویت ارسال نشده (Authorization: Bearer <token>)." } } }
      },
      "Forbidden": {
        "description": "توکن معتبر است ولی اجازه‌ی این عملیات را ندارید.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "این بخش فقط برای مالک پلتفرم در دسترس است." } } }
      },
      "NotFound": {
        "description": "منبع یافت نشد — یا اصلاً وجود ندارد، یا متعلق به کاربر دیگری است (این دو حالت عمداً یکسان جواب داده می‌شوند تا اطلاعاتی درباره‌ی داده‌ی کاربران دیگر درز نکند).",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" }, "example": { "error": "محصول یافت نشد." } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string", "description": "پیام خطای قابل‌نمایش به کاربر" } },
        "required": ["error"]
      },
      "User": {
        "type": "object",
        "description": "فیلدهای عمومی و امن کاربر — passwordHash هرگز در هیچ پاسخی برنمی‌گردد.",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "isOwner": { "type": "boolean" },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ApiKey": {
        "type": "object",
        "description": "keyHash (هش SHA-256 کل توکن) هرگز در هیچ پاسخی برنمی‌گردد؛ فقط این فیلدهای امن.",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "keyPrefix": { "type": "string", "example": "abai_live_9f6e817c..." },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "lastUsedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ApiKeyCreated": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiKey" },
          {
            "type": "object",
            "properties": {
              "token": { "type": "string", "description": "توکن کامل — فقط همین یک‌بار نمایش داده می‌شود.", "example": "abai_live_9f6e817cb2a145f0..." },
              "warning": { "type": "string" }
            }
          }
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "shortCode": { "type": "string" },
          "descriptionText": { "type": "string" },
          "isActive": { "type": "boolean" },
          "userId": { "type": "string", "nullable": true, "description": "null یعنی محصول سراسری (global) و همه‌ی کاربران می‌بینندش" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Contact": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "phone": { "type": "string", "example": "989123456789" },
          "status": { "type": "string", "enum": ["pending", "sent", "failed", "no_bale"] },
          "sentAt": { "type": "string", "format": "date-time", "nullable": true },
          "errorMessage": { "type": "string", "nullable": true }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "sender": { "type": "string", "enum": ["user", "ai", "human"] },
          "text": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Conversation": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "contactId": { "type": "string" },
          "productId": { "type": "string", "nullable": true },
          "state": { "type": "string", "enum": ["active", "completed", "abandoned"] },
          "needsHuman": { "type": "boolean" },
          "escalatedAt": { "type": "string", "format": "date-time", "nullable": true },
          "escalationReason": { "type": "string", "nullable": true },
          "contact": { "$ref": "#/components/schemas/Contact" },
          "product": { "$ref": "#/components/schemas/Product" },
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/Message" } }
        }
      },
...
      "ActivityLogEntry": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "action": { "type": "string", "example": "campaign_started" },
          "resourceType": { "type": "string", "example": "campaign" },
          "resourceId": { "type": "string", "nullable": true },
          "metadata": { "type": "string", "nullable": true, "description": "JSON string با جزئیات کمکی" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "AdminApiKeySummary": {
        "allOf": [
          { "$ref": "#/components/schemas/ApiKey" },
          {
            "type": "object",
            "properties": {
              "user": { "$ref": "#/components/schemas/User" },
              "activitySummary": { "type": "object", "additionalProperties": { "type": "integer" }, "example": { "product_created": 3, "campaign_started": 5, "contacts_imported": 2 } }
            }
          }
        ]
      },
      "AdminActivityEntry": {
        "allOf": [
          { "$ref": "#/components/schemas/ActivityLogEntry" },
          { "type": "object", "properties": { "user": { "$ref": "#/components/schemas/User" }, "apiKey": { "$ref": "#/components/schemas/ApiKey" } } }
        ]
      },
      "PaginatedAdminActivity": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/AdminActivityEntry" } },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api-keys": {
      "get": {
        "summary": "لیست کلیدهای API من",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/ApiKey" } } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "summary": "ساخت کلید API جدید",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" } } }, "example": { "name": "سرور تولید" } } }
        },
        "responses": {
          "201": { "description": "توکن کامل فقط همین یک‌بار نمایش داده می‌شود", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiKeyCreated" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/products": {
      "get": {
        "summary": "لیست محصولات من + محصولات سراسری",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "summary": "ساخت محصول جدید",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["name", "descriptionText"], "properties": { "name": { "type": "string" }, "shortCode": { "type": "string" }, "descriptionText": { "type": "string" } } },
              "example": { "name": "کرم مرطوب‌کننده", "descriptionText": "حجم ۵۰ میلی‌لیتر، قیمت ۲۵۰ هزار تومان" }
            }
          }
        },
        "responses": {
          "201": { "description": "ساخته شد", "content": { "application/json": { "schema": { "type": "object", "properties": { "product": { "$ref": "#/components/schemas/Product" } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "جزئیات یک محصول",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "product": { "$ref": "#/components/schemas/Product" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "summary": "ویرایش/فعال-غیرفعال کردن محصول (فقط مالک محصول، نه محصولات سراسری)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "name": { "type": "string" }, "descriptionText": { "type": "string" }, "isActive": { "type": "boolean" } } },
              "example": { "isActive": false }
            }
          }
        },
        "responses": {
          "200": { "description": "به‌روزرسانی شد", "content": { "application/json": { "schema": { "type": "object", "properties": { "product": { "$ref": "#/components/schemas/Product" } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/conversations/{id}/reply": {
      "post": {
        "summary": "پاسخ دستی به یک مکالمه‌ی ارجاع‌شده — مستقیماً به بازو می‌رود",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["text"], "properties": { "text": { "type": "string" } } }, "example": { "text": "سلام، بله موجوده. ارسالش رایگانه." } } }
        },
        "responses": {
          "200": { "description": "ارسال شد", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } } } } },
          "400": { "description": "متن خالی است یا مخاطب chat_id بازو ندارد", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/conversations/{id}/resolve": {
      "post": {
        "summary": "پایان ارجاع — مکالمه به AI برمی‌گردد و دوباره خودکار پاسخ می‌دهد",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "conversation": { "$ref": "#/components/schemas/Conversation" } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "summary": "[فقط مالک پلتفرم] لیست همه‌ی کاربران",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/admin/api-keys": {
      "get": {
        "summary": "[فقط مالک] همه‌ی کلیدهای API + خلاصه فعالیت هرکدام",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/AdminApiKeySummary" } } } } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/admin/activity": {
      "get": {
        "summary": "[فقط مالک] لاگ کامل فعالیت، با فیلتر و صفحه‌بندی",
        "parameters": [
          { "name": "userId", "in": "query", "schema": { "type": "string" }, "description": "فقط فعالیت این کاربر" },
          { "name": "apiKeyId", "in": "query", "schema": { "type": "string" }, "description": "فقط فعالیت این کلید API" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaginatedAdminActivity" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    }
  }
}